+++ filament/backend/include/private/backend/BackendUtils.h
@@ -28,10 +28,16 @@ namespace backend {
 size_t getFormatSize(TextureFormat format) noexcept;
 
 /**
- * For compressed texture formats, returns the number of horizontal pixels per block. Otherwise
+ * For compressed texture formats, returns the number of horizontal texels per block. Otherwise
  * returns 0.
  */
 size_t getBlockWidth(TextureFormat format) noexcept;
+
+/**
+ * For compressed texture formats, returns the number of vertical texels per block. Otherwise
+ * returns 0.
+ */
+size_t getBlockHeight(TextureFormat format) noexcept;
 
 } // namespace backend
 } // namespace filament