+++ RELEASE_NOTES.md
@@ -12,6 +12,7 @@ A new header is inserted each time a *tag* is created.
 - Fixed incorrect punctual light binning (affected Metal and Vulkan backends).
 - Fixed crash when using a Metal headless SwapChain with an Intel integrated GPU.
 - cmgen now places KTX files directly in the specified deployment folder.
+- Added support for ASTC textures on iOS with Metal backend.
 
 ## v1.4.3